fix hide button aksi

+++ view/page/donasi.php
@@ -63,7 +63,7 @@ $filter = [
                         </div>
                     </div>
                     <div class="d-flex justify-content-between grid-margin">
-                        <?php if (isAdminOrTakmir()) : ?>
+                        <?php if (isAdminOrTakmir("donasi.php")) : ?>
                             <a href="./tambah_donasi.php" class="btn btn-primary font-weight-bold text-white">Tambah Donasi</a>
                         <?php endif; ?>
 
@@ -82,7 +82,10 @@ $filter = [
                                                 <th>Nama Donatur</th>
                                                 <th>Keterangan</th>
                                                 <th>Jumlah</th>
-                                                <th>Aksi</th>
+                                                <?php if (isAdminOrTakmir()) : ?>
+                                                    <th>Aksi</th>
+                                                <?php endif; ?>
+
                                             </tr>
                                         </thead>
                                         <tbody>
@@ -98,7 +101,9 @@ $filter = [
                                                     <td></td>
                                                     <td></td>
                                                     <td></td>
-                                                    <td></td>
+                                                    <?php if (isAdminOrTakmir()) : ?>
+                                                        <td></td>
+                                                    <?php endif; ?>
                                                 </tr>
                                                 <?php
                                             } else {
@@ -110,16 +115,20 @@ $filter = [
                                                         <td><?= $val['nama_donatur'] ?></td>
                                                         <td><?= $val['ket_kasmasuk'] == "No Description" ? null : $val['ket_kasmasuk'] ?></td>
                                                         <td><?= $val['jml_donasi'] ?></td>
-                                                        <td>
-                                                            <a type="button" href="./edit_donasi.php?id=<?= $val['id_donasi'] ?>" class="btn btn-outline-secondary btn-icon-text">
-                                                                Edit
-                                                                <i class="ti-file btn-icon-append"></i>
-                                                            </a>
-                                                            <a type="button" href="./delete_donasi.php?id=<?= $val['id_donasi'] ?>" class="btn btn-outline-danger btn-icon-text">
-                                                                <i class="ti-upload btn-icon-prepend"></i>
-                                                                Hapus
-                                                            </a>
-                                                        </td>
+
+                                                        <?php if (isAdminOrTakmir()) : ?>
+                                                            <td>
+                                                                <a type="button" href="./edit_donasi.php?id=<?= $val['id_donasi'] ?>" class="btn btn-outline-secondary btn-icon-text">
+                                                                    Edit
+                                                                    <i class="ti-file btn-icon-append"></i>
+                                                                </a>
+                                                                <a type="button" href="./delete_donasi.php?id=<?= $val['id_donasi'] ?>" class="btn btn-outline-danger btn-icon-text">
+                                                                    <i class="ti-upload btn-icon-prepend"></i>
+                                                                    Hapus
+                                                                </a>
+                                                            </td>
+                                                        <?php endif; ?>
+
                                                     </tr>
                                             <?php
                                                 }
